requirements.txt
streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1
openai==1.12.0
anthropic==0.18.1
boto3==1.34.34
langchain==0.1.6
langchain-community==0.0.20
python-dotenv==1.0.1
openpyxl==3.1.2
pillow==10.2.0
pyyaml==6.0.1
